% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquis.R
\name{acquis}
\alias{acquis}
\title{Compute growth gains (acquis de croissance) at a given month.}
\usage{
acquis(X, cols, month)
}
\arguments{
\item{X}{A tibble/df with quarterly data returned by}

\item{cols}{Vector of characters, indicators for which the gain is computed.}

\item{month}{Numeric indicating the month of the quarter at which the gain is
computed.}
}
\value{

}
\description{
Should be used on quarterly data returned by \code{month_to_quarter} function.
}
