% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_diff.R
\name{add_diff}
\alias{add_diff}
\title{Add n-difference to the dataset}
\usage{
add_diff(X, exclude = NULL, n = 1L)
}
\arguments{
\item{X}{A df/tibble}

\item{exclude}{A vector of columns names indicating columns to ignore when
computing the fd. Names can be passed directly as if they were variables in
the environment - if missing defaults to \code{NULL}.}

\item{n}{A numeric indicating the order of the differencied variables.}
}
\value{
A df/tibble augmented with differencied variables.
}
\description{
Add n-difference to the dataset
}
