% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummy.R
\name{add_dummy}
\alias{add_dummy}
\title{Add dummies to the dataset}
\usage{
add_dummy(data, names, ...)
}
\arguments{
\item{data}{A df/tibble}

\item{names}{A list of characters indicating names for the dummies}

\item{...}{Logical expressions}
}
\value{
A df/tibble with additional columns with 1s if \code{conds} is \code{TRUE}
and 0s otherwise.
}
\description{
Add dummies to the dataset
}
\examples{
X <- add_dummy(X, names = list("dummy1", "dummy2"), cond1, cond2)
}
