% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_firstdiff.R
\name{add_firstdiff}
\alias{add_firstdiff}
\title{Add first-difference to the dataset}
\usage{
add_firstdiff(data, exclude = NULL)
}
\arguments{
\item{data}{A df/tibble}

\item{exclude}{A vector of characters indicating columns to ignore when
computing the fd - if missing defaults to \code{NULL}.}
}
\value{
A df/tibble augmented with fd
}
\description{
Add first-difference to the dataset
}
