% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_target.R
\name{build_target}
\alias{build_target}
\title{Build target dataset}
\usage{
build_target(
  y,
  growth_rate,
  date_freq = c("month", "quarter"),
  start = NULL,
  end = NULL
)
}
\arguments{
\item{y}{A tibble/df with two columns : date and target variable.}

\item{growth_rate}{A logical, indicates whether to compute the growth rate
or not - if missing defaults to \code{FALSE}.}

\item{date_freq}{A character \code{"month"} or \code{"quarter"}, indicates
the frequency of the date column.}

\item{start}{A character indicating the first date to keep, it must be of
the form \code{"YYYY-MM-01"} - if missing defaults to \code{NULL}.}

\item{end}{A character indicating the last date to keep, it must be of
the form \code{"YYYY-MM-01"} - if missing defaults to \code{NULL}.}
}
\value{
A tibble/df with quarterly data.
}
\description{
Build target dataset
}
