% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Compare models predictions}
\usage{
compare(...)
}
\arguments{
\item{...}{Objects of class etalonnage.}
}
\value{
A ggplot2 plot.
}
\description{
Compare models predictions
}
