% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_results.etalonnage}
\alias{export_results.etalonnage}
\title{export_results}
\usage{
\method{export_results}{etalonnage}(x, coef = TRUE, metrics = TRUE)
}
\arguments{
\item{x}{An object of class etalonnage.}

\item{coef}{A logical indicating whether to export estimated coefficients to
a csv file - if missing defaults to \code{TRUE}.}

\item{metrics}{A logical indicating whether to export some performances
metrics to a csv file - if missing defaults to \code{TRUE}.}
}
\description{
Export results to a csv file. The files are automatically named according to
the \code{name} attribute of the object or to its name if the attribute is \code{NULL}.
}
