% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_series.R
\name{extend_series}
\alias{extend_series}
\title{Extend series with the best ARIMA(p,d,q) model}
\usage{
extend_series(X, cols, n = 1)
}
\arguments{
\item{X}{A df/tibble.}

\item{cols}{A vector of characters indicating columns to extend.}

\item{n}{An integer indicating the number of samples to remove and predict.
TODO: generalize to n > 1
TODO: select cols with dplyr selection helpers}
}
\value{
A df/tibble
}
\description{
The function relies on forecast::auto.arima.
See \url{https://www.rdocumentation.org/packages/forecast/versions/8.15/topics/auto.arima}
}
