% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph.etalonnage}
\alias{graph.etalonnage}
\title{graph}
\usage{
\method{graph}{etalonnage}(
  x,
  start_graph = x$first_date,
  title = NULL,
  legend_text_size = 12,
  axis_text_size = 11,
  annotation_size = 5.5,
  date_breaks = "1 year",
  y_limits = NULL,
  annotate = FALSE,
  annotation_x = 45,
  annotation_y = -0.06
)
}
\arguments{
\item{x}{An object of class etalonnage.}

\item{start_graph}{A character indicating the first date to plot, must be of
the form \code{"YYYY-MM-01"} - if missing defaults to \code{x$first_date}.}

\item{title}{A character indicating a title for the plot - if missing
defaults to \code{NULL}.}

\item{legend_text_size}{A numeric - if missing defaults to 12.}

\item{axis_text_size}{A numeric - if missing defaults to 11.}

\item{annotation_size}{A numeric - if missing defaults to 5.5.}

\item{date_breaks}{A character indicating date breaks on the x axis of the
plot - if missing defaults to \code{"1 year"}.}

\item{y_limits}{A vector of two non-\code{NULL} numerics (ymin, ymax) - if missing
defaults to \code{NULL}.}

\item{annotate}{A logical - if missing defaults to \code{FALSE}.}

\item{annotation_x}{A numeric indicating the position of the annonation on
x-axis - if missing defaults to 45 (days).}

\item{annotation_y}{A numeric indicating the position of the annonation on
y-axis - if missing defaults to -0.06 (- 6 \%).}
}
\value{
A ggplot2 plot.
}
\description{
S3 method for class \code{etalonnage}.
}
