% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(path, date = TRUE, first_date = NULL, frequency = NULL, sheet = NULL)
}
\arguments{
\item{path}{A path to the xlsx file containing all the data.}

\item{date}{A logical indicating whether the first column of the file is
a date column. The date must be of the form \code{"YYYY-MM-01"}
\itemize{
\item if missing defaults to \code{TRUE}.
}}

\item{first_date}{A character indicating the date of observation of the first
sample when \code{date} is \code{FALSE}, must be of the form \code{"YYYY-MM-01"}
\itemize{
\item if missing defaults to \code{NULL}.
}}

\item{frequency}{A character \code{"month"} or \code{"quarter"} used to generate
dates when \code{date = FALSE} - if missing defaults to \code{NULL}.}

\item{sheet}{A character, sheet in the xlsx file - if missing defaults
to \code{NULL}.}
}
\value{
A df/tibble with a date column and loaded data
}
\description{
Load data from an xlsx file.
}
