% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{mda}
\alias{mda}
\title{mda (Mean Directional Accuracy)}
\usage{
mda(actual, predicted, lag = 1L)
}
\arguments{
\item{actual}{A numeric vector indicating the ground truth values.}

\item{predicted}{A numeric vector indicating the predicted values.}

\item{lag}{A numeric - if missing defaults to 1}
}
\value{

}
\description{
mda (Mean Directional Accuracy)
}
