% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize data}
\usage{
standardize(X, exclude = NULL, mode = c("center", "scale"))
}
\arguments{
\item{X}{A df/tibble}

\item{exclude}{A vector of columns names indicating columns to ignore when
computing the fd. Names can be passed directly as if they were variables in
the environment - if missing defaults to \code{NULL}.}

\item{scale}{A character indicating whether to scale data or only center.
Must be one of "center" or "scale".}
}
\value{
A df/tibble with standardized columns
}
\description{
Standardize data
}
