% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_data.R
\name{standardize_data}
\alias{standardize_data}
\title{Standardize data}
\usage{
standardize_data(data, exclude = NULL, scale = FALSE)
}
\arguments{
\item{data}{A df/tibble}

\item{exclude}{A vector of characters indicating columns to ignore when
standardizing - if missing defaults to \code{NULL}.}

\item{scale}{A logical indicating whether to scale data or only center - if
missing defaults to \verb{FALSE``.}}
}
\value{
A df/tibble with standardized columns
}
\description{
Standardize data
}
