% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_growth_rate.R
\name{to_growth_rate}
\alias{to_growth_rate}
\title{Convert series to growth rate}
\usage{
to_growth_rate(X, exclude = NULL, inplace = TRUE)
}
\arguments{
\item{X}{A df/tibble}

\item{exclude}{A vector of columns names indicating columns to ignore when
computing the fd. Names can be passed directly as if they were variables in
the environment - if missing defaults to \code{NULL}.}

\item{inplace}{A logical indicating whether to replace values or add values
as new columns to \code{X} - if missing defaults to \code{TRUE}.}
}
\value{
A df/tibble augmented with converted variables.
}
\description{
Convert series to growth rate
}
