% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etalonnage.R
\name{etalonnage}
\alias{etalonnage}
\title{etalonnage}
\usage{
etalonnage(
  name,
  X,
  y,
  model_type,
  train_index,
  test_index,
  scale = c(NULL, "center", "scale"),
  seed = 313,
  ...
)
}
\arguments{
\item{name}{A character indicating a name for the analysis.}

\item{...}{}
}
\value{
An object of the S3 class etalonnage.
}
\description{
etalonnage
}
