% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_series.R
\name{extend_series}
\alias{extend_series}
\title{Extend series}
\usage{
extend_series(data, cols, n = 1)
}
\arguments{
\item{data}{A df/tibble}

\item{cols}{A vector of characters indicating columns to extend}

\item{n}{An integer indicating the number of samples to remove and predict
TODO: generalize to n > 1
TODO: select cols with dplyr selection helpers}
}
\value{
A df/tibble
}
\description{
Extend series
}
