% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph.etalonnage}
\alias{graph.etalonnage}
\title{Title}
\usage{
\method{graph}{etalonnage}(x, start, title = NULL, legend_text_size = 12, axis_text_size = 11, ...)
}
\arguments{
\item{start}{A character indicating the date of observation of the first
sample, must be of the form "YYYY-MM-01"}

\item{...}{}
}
\value{

}
\description{
Title
}
