% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_data.R
\name{standardize_data}
\alias{standardize_data}
\title{Standardize data}
\usage{
standardize_data(X, exclude = NULL, scale = FALSE)
}
\arguments{
\item{X}{A df/tibble}

\item{exclude}{A vector of columns names indicating columns to ignore when
computing the fd. Names can be passed directly as if they were variables in
the environment - if missing defaults to \code{NULL}.}

\item{scale}{A logical indicating whether to scale data or only center - if
missing defaults to \verb{FALSE``.}}
}
\value{
A df/tibble with standardized columns
}
\description{
Standardize data
}
