% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_index.R
\name{train_test_index}
\alias{train_test_index}
\title{train_test_index}
\usage{
train_test_index(X, date1, date2, horizon = 1)
}
\arguments{
\item{X}{A df/tibble containing the predictors.}

\item{date1}{A character indicating the starting date of the initial window.
It must be of the form "YYYY-MM-01".}

\item{date2}{A character indicating the last date of the initial window.
It must be of the form "YYYY-MM-01".}

\item{horizon}{A integer indicating the number of consecutive values in test
set sample - if missing defaults to 1.}
}
\value{
A list with 2 items: one for training indexes, one for test indexes.
}
\description{
train_test_index
}
